Conflicts:
	definitions/mip_install_perl.sh

+++ definitions/mip_install_perl.sh
@@ -52,15 +52,21 @@ error() {
 ## Enable trap for signal(s) ERR
 trap '$(error "$previous_command" "$?")' ERR
 
+
 ## Create or install conda env
-if [ "$EXISTING_ENV" = true ]
+if [ "$EXISTING_ENV" = true ] && [ -d ${CONDA_PATH}/envs/${ENV_NAME} ]
 then
   conda install --name "$ENV_NAME" --yes  -c conda-forge libgcc-ng gxx_linux-64
+elif [ -d ${CONDA_PATH}/envs/${ENV_NAME} ]
+then
+    echo "Environment already exists. Please supply flag -x to command"
+    exit 1
 else
   conda create --name "$ENV_NAME" --yes  -c conda-forge libgcc-ng gxx_linux-64
 fi
 
-conda install --name "$ENV_NAME" --yes -c bioconda -c conda-forge perl=5.26 perl-app-cpanminus perl-log-log4perl perl-moosex-app perl-file-copy-recursive perl-timedate
+conda install --name "$ENV_NAME" --yes -c bioconda -c conda-forge perl=5.26 perl-app-cpanminus perl-log-log4perl perl-moosex-app perl-file-copy-recursive perl-timedate  perl-set-intervaltree
+conda install --name "$ENV_NAME" --yes -c bioconda -c conda-forge perl=5.26 perl-app-cpanminus perl-log-log4perl perl-moosex-app perl-file-copy-recursive perl-set-intervaltree
 
 ## Source conda
 source "$CONDA_PATH"/etc/profile.d/conda.sh
